refactor: split ResourceCardDescription into editable and non-editable components
