+++ src/Components/ResourceList/Card/ResourceCardDescription.tsx
@@ -1,14 +1,9 @@
 // Libraries
-import React, {Component, KeyboardEvent, ChangeEvent} from 'react'
+import React, {Component} from 'react'
 import classnames from 'classnames'
 
-// Components
-import {Input} from '../../Inputs/Input'
-import {Icon} from '../../Icon/Icon'
-import {ClickOutside} from '../../ClickOutside/ClickOutside'
-
 // Types
-import {StandardProps, ComponentSize, IconFont} from '../../../Types'
+import {StandardProps} from '../../../Types'
 
 // Styles
 import './ResourceCardDescription.scss'
@@ -16,56 +11,22 @@ import './ResourceCardDescription.scss'
 interface Props extends StandardProps {
   /** Text to display in description */
   description: string
-  /** Placeholder text to display in input during editing */
-  placeholder: string
-  /** Called when user hits enter or blurs the input  */
-  onUpdate: (description: string) => void
 }
 
-interface State {
-  isEditing: boolean
-  workingDescription: string
-}
-
-export class ResourceCardDescription extends Component<Props, State> {
+export class ResourceCardDescription extends Component<Props> {
   public static readonly displayName = 'ResourceCardDescription'
 
   public static defaultProps = {
     testID: 'resource-list--description',
-    placeholder: '',
-  }
-
-  constructor(props: Props) {
-    super(props)
-
-    this.state = {
-      isEditing: false,
-      workingDescription: props.description,
-    }
   }
 
   public render() {
     const {description, testID, id} = this.props
-    const {isEditing} = this.state
-
-    if (isEditing) {
-      return (
-        <div className={this.className} data-testid={testID} id={id}>
-          <ClickOutside onClickOutside={this.handleStopEditing}>
-            {this.input}
-          </ClickOutside>
-        </div>
-      )
-    }
 
     return (
-      <div className={this.className} data-testid={testID}>
-        <div
-          className={this.previewClassName}
-          onClick={this.handleStartEditing}
-        >
+      <div className={this.className} data-testid={testID} id={id}>
+        <div className={this.previewClassName}>
           {description || 'No description'}
-          <Icon glyph={IconFont.Pencil} />
         </div>
       </div>
     )
@@ -75,62 +36,6 @@ export class ResourceCardDescription extends Component<Props, State> {
     const {className} = this.props
 
     return classnames('resource-description', {[`${className}`]: className})
-  }
-
-  private get input(): JSX.Element {
-    const {placeholder} = this.props
-    const {workingDescription} = this.state
-
-    return (
-      <Input
-        size={ComponentSize.ExtraSmall}
-        autoFocus={true}
-        spellCheck={false}
-        placeholder={placeholder}
-        onFocus={this.handleInputFocus}
-        onChange={this.handleInputChange}
-        onKeyDown={this.handleKeyDown}
-        className="resource-description--input"
-        value={workingDescription}
-      />
-    )
-  }
-
-  private handleStartEditing = (): void => {
-    this.setState({isEditing: true})
-  }
-
-  private handleStopEditing = async (): Promise<void> => {
-    const {workingDescription} = this.state
-    const {onUpdate} = this.props
-
-    await onUpdate(workingDescription)
-
-    this.setState({isEditing: false})
-  }
-
-  private handleInputChange = (e: ChangeEvent<HTMLInputElement>): void => {
-    this.setState({workingDescription: e.target.value})
-  }
-
-  private handleKeyDown = async (
-    e: KeyboardEvent<HTMLInputElement>
-  ): Promise<void> => {
-    const {onUpdate, description} = this.props
-    const {workingDescription} = this.state
-
-    if (e.key === 'Enter') {
-      await onUpdate(workingDescription)
-      this.setState({isEditing: false})
-    }
-
-    if (e.key === 'Escape') {
-      this.setState({isEditing: false, workingDescription: description})
-    }
-  }
-
-  private handleInputFocus = (e: ChangeEvent<HTMLInputElement>): void => {
-    e.currentTarget.select()
   }
 
   private get previewClassName(): string {
