+++ src/index.ts
@@ -81,6 +81,7 @@ export * from './Components/ResourceList/List/ResourceListHeader'
 export * from './Components/ResourceList/List/ResourceListSorter'
 export * from './Components/ResourceList/Card/ResourceCard'
 export * from './Components/ResourceList/Card/ResourceCardDescription'
+export * from './Components/ResourceList/Card/ResourceCardEditableDescription'
 export * from './Components/ResourceList/Card/ResourceCardName'
 export * from './Components/ResourceList/Card/ResourceCardEditableName'
 export * from './Components/Table/Table'
